+++ src/git-annex-finder/MenuCommandTypeTag.swift
@@ -1,0 +1,15 @@
+//
+//  MenuCommandTypeTag.swift
+//  git-annex-finder
+//
+//  Created by Andrew Ringler on 5/1/18.
+//  Copyright © 2018 Andrew Ringler. All rights reserved.
+//
+
+import Foundation
+
+enum MenuCommandTypeTag: Int {
+    case selectedItems = 1
+    case target = 2
+    case repoOfTarget = 3
+}